rename find_face to identify_face, add find face
